Move Environment to entities
