+++ src/routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Environment;
+use Entities\Environment;
 
 /*
 |--------------------------------------------------------------------------
